historical drafts: another catalog sketch!
(surely this is the last one!)

Another may draft.  Nevermind the bullocks that this isn't actually a valid yaml document.

This one is gradually giving up more and more on using hashes in any part of the document tree except for the final leafs, and giving human names more and more role.  Today, this continues to seem valid (and the 'baw'-like comment at the bottom somewhat misplaced): we need to give all steps a name, because we actually *do not* typically want them to collapse if they *happen* to converge, and the human names assigned to nodes is a good, meaningful, simple and comprehensible way to do so.  (Plus all the standard handwave that we're SOL for convergence from the first step of these documents anyway, so might as well go all in on the flexibility that affords.)

Signed-off-by: Eric Myhre <hash@exultant.us>