historical drafts: now with structs!
Yesterday's attempts to put pen to structs.

Explicitly now chosing the idea that there will be three pieces to a name, and that will go directly to a ware ID -- at no point do you get thrown a set of wares and told to choose based on metadata, as some earlier drafts would have held.

Signed-off-by: Eric Myhre <hash@exultant.us>