api: updated example.  now 'step' system.  docs.
Many, many docs.  And a couple odd comments at the end that are notes-to-future-self.

With this, our first steps on making 'verbs' lines up with our structs nicely again; formulas have a clear *place* in this iteration of the drafts and that seems like an obvious success score point; and we've finally taken a clear stance on where indeed hashes are vendored and snapshotted into the process here, which is critical to answer and answer well.  This feels good.  We're close, here.

Signed-off-by: Eric Myhre <hash@exultant.us>