api: refmt atlas for ReleaseItemID; test fixtures.
Instantly bumping refmt version because I fatfingered getting the latest when initializing the link.  But we'll see this link bump a *lot*; I'm going to be using hitch here as leverage to force me to fix all the todos in refmt (while writing test coverage in refmt like a mature adult, of course).

Signed-off-by: Eric Myhre <hash@exultant.us>