api: large (large) serial fixture test.
Consumed and removed previous example var, porting it to this test.  Added remaining snippet of atlas generation to make all fly.

It's certainly a verbose structure.  Some room for trimming, perhaps.  (In particular that runRecord.Metadata field should be omitempty, etc, but we're not concerning ourselves with such surface detail at this moment.)

It seems very likely that several parts of this document should be split into separate files in our default on-filesystem layout.  In particular, for the formula component to be expressed in its own file seems like a reasonable core choice.  We'll likely follow that path later; keeping it as it stands for this example at present.

Ha!  And I just noticed I blithely admitted 'uID' to the fixture.  Yeah, ok, well, no heuristic or default is perfect.  We'll expect to have explicit atlases for many of these structures, I think, in the final versions.

Neat: in this fixture you can see demonstrated our use of reified structs for the release ID -- e.g. `ReleaseItemID{"hub.repeatr.io/base", "2017-05-01", "linux-amd64"}` -- while the serial form contains the translated simple-to-handle string form: `"hub.repeatr.io/base:2017-05-01:linux-amd64"`.

Signed-off-by: Eric Myhre <hash@exultant.us>