rdef: define formula type and atlases.
This structure is roughly in keeping with the current v0.15ish repeatr formula format, but not exactly.  We're effectively defining the "v200" formula format here: shoving in all of our desired but breaking changes at once.  Hitch code will thus always be written against the v200 format as we develop it.  We'll likely build a munger that turns these v200 formulas into v0.15 formulas to ease migration.

FormulaAction is still todo.  Some potential choices to make about output slots in the future as well.

Test fixtures updated: since the Formula type is now a struct rather than a map, the default atlas defines serialization order as the struct field source order rather than the alphasort which maps default to, and as a result "action" is now in the middle rather than the top.

Signed-off-by: Eric Myhre <hash@exultant.us>