cli: incremental progress.
Fix some typos in the subcommand rigging; validated that required positional args actually do work great! (yay!); changed the switch default case to be a panic because it's now only reachable if you *do* hit some subcommand (otherwise the error is emitted by kingpin earlier on), so it's a bug of ours if that's not handleable.

Signed-off-by: Eric Myhre <hash@exultant.us>