core: add data fields to stage controller.
Outline sync and load.  Not sure handing *down* the files like that makes much sense in context -- we're likely using multiple files, later -- but rolling with it for now.

May also be useful to make different type rather than reusing api.Catalog, because maintaining the invarient of a one-element slice doesn't seem terribly fun, but again, rolling with it for now.

Signed-off-by: Eric Myhre <hash@exultant.us>