lib/errcat: embark on a new world of error science
This is probably re-usable and worthy of its own package, but it's a new idea, and I'm just going to incubate it here, in-tree, for now.

The aim is to make local error handling *possible*, and do it *without* resorting to arbitrarily-deep "wrapping"/"cause" approaches -- because writing a linked list traverser is the kind of complexity that I LEAST ought to be engaging with whilst already in the middle of handling an unhappy state.

Another success heuristic here is that error handling statements should never import any symbols from packages that aren't already imported.  In other words, if the error handling code for `somepkg.Func()` refers to any packages *other* than `somepkg`, that's an unacceptable code smell and leakage of implementation details, and we ought to consider it a bug.  (This is something I have *a lot* of problems with in other major projects.  And no, refactoring them all into a shared errors package is not a solution: it's a junkdrawer, and another code smell of its own.)  The errcat design is intended to make that easy to get right, by making it easy to reassign category information as errors propagate.

We'll see how this goes.

Signed-off-by: Eric Myhre <hash@exultant.us>