lib/errcat,core: refactor release-add-label.
Seealso the previous commit, which introduces errcat, and describes some of the motivations in the commit message.

This diff shows roughly all the things we wanted to go right... going right: the somepkg.Func call is followed by a switch which entirely and exclusively refers to somepkg.Consts; the returned values are effectively a recategorization (and we'll probably rewrite this and the other core package code to make that more literally true in future commits).

The odd detail of the "err2" variable rename is because the existing "err" variable handle refers to an "error" type.  This would go away when we refactor the entire method to consistently use errcat methods, but is still a bit of an irritant.

Signed-off-by: Eric Myhre <hash@exultant.us>