core: fix regression in release-start err handling
It's the typed-nil issue: it strikes again.  Apparently by ever threatening to assign a maybe-value of a pointer type which is clearly never not null... if a function defn (for example) gives it type-yness... this actually may have TWO outcomes:

- If the assignment in the calling context is into a slot of the same pointer type, *it's fine*: it remains an untyped nil, like you'd expect.
- If the assignment in the calling context is into a more general interface, *your untyped nil becomes typed*.  I, for one, did not expect this.

Even *knowing about* the typed-nil concept, this still surprises, confounds, and irritates me.  There's a "reason".  But damn if this isn't an unergonomic source of... well, screaming.  Gnashing of teeth, if you want to be biblical.  It sucks.

Next commit: I'm adding helper methods do make the most important operations available without direct access the the errcat.Error type, and refactoring everything back to consistently use `error` again... this is the only sure way to avoid typed-nil hell, it seems.

Signed-off-by: Eric Myhre <hash@exultant.us>