lib/testutil: it's time...
Bring in some near-copypasta for tempdir help for use in testing.

Writing almost any interesting tests for hitch involves conceding use of the filesystem.  Typically it's good to be wary of tests that have sideeffects or inputs from the filesystem (they're at least usually "medium"/not-"unit" anymore), but given the way hitch works can be effectively described an op-log whose natural resting state is to be persisted to disk... we reach the point of needing this category of testing very very quickly.

Signed-off-by: Eric Myhre <hash@exultant.us>