all: consistently use "item name", not "label".
The word "label" is okay to use in full sentences describing what's going on, but we're expunging it from the code.

The result is much more consistent: with CatalogName and ReleaseName we have ItemName.

I have no idea why I let that inconsistency fester this long.

The command `hitch release add-label` is now also `hitch release add-item`.  Here, it's almost more debatable: if you were using the form of the command that takes a "wire" rather than an explicit WareID, then "label" might actually be a valid verb for what's being done.  But when it IS an explicit WareID, it's definitely "adding" an item rather than merely "labelling", because it's never been mentioned before.

Signed-off-by: Eric Myhre <hash@exultant.us>