core: catalog-create cmd! Now required to release.
Test coverage is now up there and feeling good!  Getting called on lots of B.S. immediately.

Oh, and now the `hitch release commit` tests work!  And, the demo script had to be updated to have a catalog-create step before the release staging starts, as did all the other tests, of course.

Starting to experience some serious redundancy on some of this serialization and basic file IO code, but we're going to defer a DRY'ing refactor on that until it crops up at least one more time, so we don't accidentally over-focus on the wrong bits.

Signed-off-by: Eric Myhre <hash@exultant.us>