core: begin draft of show command.
I'm considering how many variations on this command we might need.  It's... probably going to be quite a few.

Specifically, I'm a little worried about how the most precise form -- when you give it the full three-tuple, all the way down to item -- is going to behave.  It seems to me that people will most often want that form to emit the wareID, bare.  But the other forms of the command emit json objects!  So, it would be *consistent* to emit the wareID as a quoted string by default.  Urgh.  We'll probably make a fairly arbitrary choice for now, and revist later.  (It's amusing to note that switching the default serial format to yaml, or something close to it, would quietly fix the problem, since the quotes would be missing then anyway.)

Also, I'm losing patience with the amount of redundancy in the {"find db", "load catalog",...} bits of code.  It's time to DRY up some of this, coming up next...

Signed-off-by: Eric Myhre <hash@exultant.us>