core: show command working for catalogs!
This isn't much more than you could do with `cat` at the moment, realistically, but it's following a good known flow, so we'll be able to enhance it as necessary in the future, add more sanity checks, more output formatter options, and so on.  And it gives you a clear exit code on things like "not found"!  And so on.

Test coverage: currently missing.  These inspection commands, particularly the ones with broad results, are going to require a new pattern of assertion.  I'm thinking just asserting on some expected substrings :/ because it doesn't seem at all useful to parse the full output back *again*.  The tests for show commands like this, particularly this one with its total lack of filtering, are pretty much smoke tests that output is wired up remotely correctly; they don't need to go deep.

We've started to extract some shared functions, like the catalog loader.  They're still using error returns, and thus end up being four-liner calls rather than one.  I continue to be tempted to write those things to use panics as package-internal non-local returns.

Signed-off-by: Eric Myhre <hash@exultant.us>