core: nonbreaking refactor show internals.
Pretty sure this also added the appropriate behaviors for showing a single item, as well, but haven't written the tests for that yet, so no official comment.

This turns out to be one of the rare situations where despite the logical flow being "switch"-like, a series of early return checks makes a lot simpler code.  The way in which each of the "cases" needs to load slightly more info than the last just didn't jive with a switch/fallthrough flow.

Signed-off-by: Eric Myhre <hash@exultant.us>