api: more accepting WareID parser.
More colons are fine.  It's just the first one that counts for the purposes of type-vs-hash separation.

We don't currently have any examples of supporting WareIDs like "git:0239ff340:path/within/tree.ext"... but there's nothing to say we couldn't in the future.  (IPFS is actively supportive of this kind of navigation, as I understand.  Though that's sort of a poor example because I also don't think they use the colon to indicate for it.)

Signed-off-by: Eric Myhre <hash@exultant.us>